refactor #255: 스케줄러 관련 클래스 위치 수정
