+++ backEnd/src/main/java/com/quiz/ourclass/domain/challenge/service/ChallengeServiceImpl.java
@@ -28,11 +28,11 @@ import com.quiz.ourclass.domain.organization.entity.MemberOrganization;
 import com.quiz.ourclass.domain.organization.entity.Organization;
 import com.quiz.ourclass.domain.organization.repository.MemberOrganizationRepository;
 import com.quiz.ourclass.domain.organization.repository.OrganizationRepository;
-import com.quiz.ourclass.global.config.scheduler.SchedulingService;
 import com.quiz.ourclass.global.exception.ErrorCode;
 import com.quiz.ourclass.global.exception.GlobalException;
 import com.quiz.ourclass.global.util.AwsS3ObjectStorage;
 import com.quiz.ourclass.global.util.UserAccessUtil;
+import com.quiz.ourclass.global.util.scheduler.SchedulingService;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
